\name{huffmanr-package}
\alias{huffmanr-package}
\alias{huffmanr}
\docType{package}
\title{
Huffman codes for R
}
\description{
Build Huffman codes in R.
}
\details{
\tabular{ll}{
Package: \tab huffmanr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-11-17\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Brad Skaggs

Maintainer: Brad Skaggs <bskaggs@acm.org>
}
\references{
~~ Literature or other references for background information ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
